pyproject.toml
[project]
name = "wave-mcp"
version = "0.1.0"
description = "MCP server for viewing and parsing VCD waveform files"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.0.0",
    "vcdvcd>=2.1",
]

[dependency-groups]
dev = [
    "pyvcd>=0.4.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[project.scripts]
wave-mcp = "mcp_server:main"
